Two major changes in this commit, please read bellow
1- Better management for sessions and errors
The script now gracefuly fails and try to send data to prometheus
regarding its issues.

2- Use labels for prometheus data instead of multiple indexes
Instead of having 4-5 counters only for ec2 instances events, there is
now only one, and we're using labels in order to allow a distinction.
This also allows to push ec2 instance_id in addition for better
troubleshooting and actions.